separated build_dog in different files to increase readability; added a floor to the base xml; overall clean-up of the file
