+++ dm_control/suite/dog_assets/BONEXiphoid_cartilage.stl
